Applying patch provided by Stjepan Gros to eliminate compiler warnings.
* openvasd/oval_plugins.c (oval_plugin_add, ovaldi_launch): Minor
reformattings, changes to eliminate compiler warnings.

* openvasd/otp_1_0.c (otp_1_0_server_send_certificates): Casting to
eliminate compiler warning.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-server@1723 423fd1db-d629-0410-8442-d21db03e70f4

+++ openvasd/otp_1_0.c
@@ -97,7 +97,7 @@ void otp_1_0_server_send_certificates(struct arglist* globals)
       cert_list_elem = g_slist_next(cert_list_elem);
       auth_printf(globals, "%s <|> %s <|> %s <|> %d <|> %s\n", cert->fpr,
                               cert->ownername, trustlevel,
-                              strlen(cert->full_public_key),
+                              (int)strlen(cert->full_public_key),
                               cert->full_public_key);
       // Release each element
       openvas_certificate_free(cert);
